requirements.txt
PyPDF2
pymupdf
Pillow
pdfplumber
pytesseract
camelot-py[cv]
pandas